moving tests to BatchingDSLSpec
